Renamed files to match naming conventions. Deleted old code that was no longer being used.
