+++ src/main/java/display_use_case/DisplayInputBoundary.java
@@ -1,4 +1,4 @@
-package DisplayUseCase;
+package display_use_case;
 import Screens.InventoryViewModel;
 
 public interface DisplayInputBoundary {
